move coordinator to separate file
